Implemented new project structure
Got rid of Edges, Nodes, and Graphs files as inheritance was is garbage
in cpp. Now we are just going to have seperate classes for each of these
things. For now I'm only focussing on Undirected graphs until I get to a
point were I want to deal with directed ones.